the version that treats xml

+++ glParseur.py
@@ -6,21 +6,27 @@ import os.path  #le path du program
 import shutil   #pour suprimmer récursivement
 
 def filtre(src,dst,element):
-		titre = element.replace('.txt','').replace('_',' ') 		# on remplace les underscore par des espaces et on enleve l'extension du fichier 
-		dst.write("Nom du pdf : "+titre+"\n")				 		# on ecrit le nom du pdf dans le fichier de destination (premiere ligne)					
+        # on remplace les underscore par des espaces et on enleve l'extension du fichier 
+		titre = element.replace('.txt','').replace('_',' ')
+        # on ecrit le nom du pdf dans le fichier de destination (premiere ligne) 		
+		dst.write("Nom du pdf : "+titre+"\n")				 							
 		for i in range(2000,2019) :								
 			if str(i) in titre :
 				annee = str(i)	
-		dst.write("Titre : "+  titre.split(annee)[1] + "\n")  		# on ecrit le titre du pdf dans le fichier de destination (deuxieme ligne)
-		txt = src.read()											# lecture du fichier dans une variable string
-		if "ABSTRACT" in txt :										# blocs de conditions pour identifier le début du résumé  
+        # on ecrit le titre du pdf dans le fichier de destination (deuxieme ligne)
+		dst.write("Titre : "+  titre.split(annee)[1] + "\n") 
+        # lecture du fichier dans une variable string 		
+		txt = src.read()
+        # blocs de conditions pour identifier le début du résumé  											
+		if "ABSTRACT" in txt :										
 			debut = txt.split("ABSTRACT",1)						
 								
 		elif "Abstract" in txt :
 			debut = txt.split("Abstract",1)
 		
 		fin="1"
-		if "Keywords" in txt :										# blocs de conditions pour identifier la fin du résumé
+        # blocs de conditions pour identifier la fin du résumé
+		if "Keywords" in txt :										
 			fin="Keywords"
 			
 		elif "Index" in txt :
@@ -29,7 +35,8 @@ def filtre(src,dst,element):
 		a1 = debut[1].split(fin)
 		a2 = a1[0].split("\n")
 		dst.write("Resumé : ")
-		for i in range(0,len(a2)) :									# ecriture du résumé dans le fichier de destination sur une seule ligne (troisieme ligne)
+        # ecriture du résumé dans le fichier de destination sur une seule ligne (troisieme ligne)
+		for i in range(0,len(a2)) :									
 			dst.write(a2[i])
 		
 
@@ -72,7 +79,7 @@ def pdf(directoryPath):
 
             # renommer le fichier 
             os.rename("{0}/{1}".format(directoryPath,fileName), "{0}/{1}".format(directoryPath,newFileName))
-            # créer la commande qui permet de 
+            # créer la commande qui permet de faire la conversion
             pdfToTextCommand = "pdftotext -f 1 {1}/{2} {1}/tmp/{3}.txt".format(os.getcwd(),directoryPath, newFileName, fileNameWithouExt)
             # Executer la commande de conversion
             os.system(pdfToTextCommand)
@@ -98,17 +105,23 @@ print("***************************************" + bcolors.ENDC)
 # S'assurer que notre programme reçois le bon nombre d'argument 
 # le premier argument est le nom de notre script python
 # le deuxième argument est le répértoire content l'ensemble des fichiers pdf à convertir.
-if len(sys.argv) != 2: 
+# le troisiéme argument est pour choisir le type de sortie soit txt ou xml
+
+if len(sys.argv) != 3: 
     print(bcolors.FAIL + "Ooops nombre d'arguments incorrect" + bcolors.ENDC)
     print("Merci de passer en paramètre le dossier contenant les fichier pds à convertir.")
-    print("Exemple: " + bcolors.OKGREEN+"python3 gl.py chemin_vers_le_dossier" + bcolors.ENDC)
+    print("et le type de sortie :")
+    print(bcolors.OKBLUE+"soit -t pour un fichier en format texte ."+bcolors.ENDC)
+    print(bcolors.OKBLUE+"soit -x pour un fichier en format xml."+ bcolors.ENDC)
+    print("Exemple 1: " + bcolors.OKGREEN+"python3 gl.py chemin_vers_le_dossier -t pour un fichier TXT" + bcolors.ENDC)
+    print("Exemple 2: " + bcolors.OKGREEN+"python3 gl.py chemin_vers_le_dossier -t pour un fichier XML" + bcolors.ENDC)
     sys.exit(2)
 else:
     # Récupérer le répértoire courant ( cwd : current working directory )
     current = os.getcwd()
     directory = sys.argv[1]
 
-    # Terminer le programme si le l'argument passé en paramètre n'existe pas
+    # Terminer le programme si l'argument passé en paramètre n'existe pas
     if not os.path.exists(directory):
         print(bcolors.FAIL + "L'argument passé en paramètre n'existe pas." + bcolors.ENDC)
         sys.exit(2)
@@ -117,10 +130,19 @@ else:
     if not os.path.isdir(directory):
         print(bcolors.FAIL + "L'argument passé en paramètre n'est pas un dossier." + bcolors.ENDC)
         sys.exit(2)
-    
-    # Début de la conversion
-    print ("Conversion des fichier du répértoire " + directory)
-    pdf(directory)	
-    transmog(directory)
-    t = "{}/tmp".format(directory)
-    shutil.rmtree(t)   
+    # Verifier si le type de sortie est égale a txt
+    if sys.argv[2] == '-t':
+        # Début de la conversion
+        print("Conversion pdf to txt")
+        print ("Conversion des fichier du répértoire " + directory)
+        pdf(directory)	
+        transmog(directory)
+        t = "{}/tmp".format(directory)
+        shutil.rmtree(t) 
+    # Verifier si le type de sortie est égale a xml
+    #elif sys.argv[2] == '-x':
+        #...
+    # Terminer le programme si l'argument de type de sortie n'égale ni à txt ni à xml 
+    else :
+        print(bcolors.FAIL + "Ooops le type de sortie est inconnue" + bcolors.ENDC)
+        sys.exit(2)
